+++ CHANGELOG.md
@@ -118,3 +118,4 @@
 
 - Update to NeoForge 21.0.166
 - Fixed the Scraper not working on the Ender Dragon (#15)
+- Made it so that removing Overclockers from the Advanced Incubator will reset its progress